% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipriarmsdata.R
\docType{data}
\name{sipri_cat}
\alias{sipri_cat}
\title{Exports by Weapons Category}
\format{
A data frame with 12115 rows and 4 variables:
\describe{
\item{Export_From}{The exporting country}
\item{Weapon_Category}{The Category of Weapon}
\item{Year}{Year}
\item{TIV}{Trade Indicator Values, in US Dollars}
...
}
}
\source{
\url{https://www.sipri.org/databases/armstransfers}
}
\usage{
sipri_cat
}
\description{
Information on all transfers of major conventional weapons from 1950 to
the most recent full calendar year
}
\keyword{datasets}
