% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipriarmsdata.R
\docType{data}
\name{sipritiv}
\alias{sipritiv}
\title{Exports by Destination}
\format{
A data frame with 30628 rows and 4 variables:
\describe{
\item{Export_From}{The exporting country}
\item{Import_To}{The importing country}
\item{Year}{Year}
\item{TIV}{Trade Indicator Values, in US Dollars}
...
}
}
\source{
\url{https://www.sipri.org/databases/armstransfers}
}
\usage{
sipritiv
}
\description{
Information on all transfers of major conventional weapons from 1950 to
the most recent full calendar year
}
\keyword{datasets}
